Add Postgres init & DB dump; update workflows
Add PostgreSQL service to docker-compose with volume and init scripts, and include a full ACS database dump (docker/init/01_init_acsdb.sql) to seed schema and demo data. Introduce NearestNodeFinder and several Elsa workflows (AMR connected/disconnected, vehicle location change, rail location workflow) plus related elsa migration updates. Update DB context, DAOs, node/path models and DTOs, MQTT interface/manager, manager/service implementations, and UI viewmodels/controls to support new node handling and workflow integration. Also add VehicleState docs and minor MQTT docs/UI adjustments.

+++ src/ACS/ACS.Communication/Http/Models/NodeDto.cs
@@ -4,8 +4,8 @@ namespace ACS.Communication.Http.Models
     {
         public string Id { get; set; }
         public string Type { get; set; }
-        public int Xpos { get; set; }
-        public int Ypos { get; set; }
-        public int Zpos { get; set; }
+        public double Xpos { get; set; }
+        public double Ypos { get; set; }
+        public double Zpos { get; set; }
     }
 }
