+++ src/ACS/ACS.UI/Models/NodeDto.cs
@@ -4,7 +4,7 @@ public class NodeDto
 {
     public string Id { get; set; }
     public string Type { get; set; }
-    public int Xpos { get; set; }
-    public int Ypos { get; set; }
-    public int Zpos { get; set; }
+    public double Xpos { get; set; }
+    public double Ypos { get; set; }
+    public double Zpos { get; set; }
 }
